+++ tests/testthat/test-read-data.R
@@ -35,7 +35,7 @@ test_that("can read iso3 from geojson", {
 
 test_that("can read country from PJNZ", {
   pjnz <- file_object(file.path("testdata", "Botswana2018.PJNZ"))
-  expect_equal(read_country(pjnz), "Botswana")
+  expect_equal(read_country(pjnz$path), "Botswana")
 })
 
 test_that("geojson read applies can use cache", {
@@ -44,4 +44,16 @@ test_that("geojson read applies can use cache", {
   value <- hintr_geojson_read(path, cache)
   expect_equal(cache$list("geojson"), path$hash)
   expect_identical(hintr_geojson_read(path, cache), value)
+})
+
+test_that("can read spectrum region code from geojson", {
+  shape <- file_object(file.path("testdata", "malawi.geojson"))
+  region_codes <- read_geojson_spectrum_region_codes(shape)
+  expect_equal(region_codes, 0)
+
+  skip_if_sensitive_data_missing()
+  shape <- file_object(file.path("testdata", "sensitive", "ZMB", "data",
+                     "zmb_areas.geojson"))
+  region_codes <- read_geojson_spectrum_region_codes(shape)
+  expect_equal(region_codes, c(10, 11, 12, 13, 14, 15, 16, 17, 18, 19))
 })