+++ tests/testthat/test-validate-inputs.R
@@ -6,15 +6,11 @@ test_that("PJNZ can be validated and return data", {
                list(data = list(country = scalar("Botswana"),
                                 iso3 = scalar("BWA")),
                     filters = scalar(NA)))
-  mock_read_country <- mockery::mock("GBR")
-  with_mock("hintr:::read_country" = mock_read_country, {
-    expect_error(do_validate_pjnz(pjnz), "Invalid country")
-  })
 })
 
 test_that("country can be read", {
   pjnz <- file_object(file.path("testdata", "Botswana2018.PJNZ"))
-  expect_equal(read_country(pjnz), "Botswana")
+  expect_equal(read_country(pjnz$path), "Botswana")
 })
 
 test_that("assert fails if more than once country in json", {
@@ -26,17 +22,6 @@ test_that("assert fails if more than once country in json", {
   json$features[[1]]$properties$area_id <- "AGO"
   expect_error(assert_single_country(json, "shape"),
                "Shape file contains regions for more than one country. Got countries AGO, MWI.")
-})
-
-test_that("assert fails if a feature is missing an area id", {
-  shape <- file_object(file.path("testdata", "malawi.geojson"))
-  json <- hintr_geojson_read(shape)
-  expect_true(assert_area_id_exists(json))
-
-  ## Remove an ID for testing
-  json$features[[1]]$properties$area_id <- NULL
-  expect_error(assert_area_id_exists(json),
-               "Shape file does not contain an area ID for each region. Missing ID for 1 feature.")
 })
 
 test_that("do_validate_shape validates shape and returns geojson as list", {